requirements.txt
numpy==1.13.3
pyrealsense2==2.18.1.549
PySide2==5.12.1
pypng==0.0.19
subprocess32==3.5.3
ruamel.yaml==0.15.89
argparse==1.1
autolab-core==0.0.1
scipy==1.0.0
matplotlib==2.1.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
scikit-image==0.13.1
scikit-learn==0.19.0
Pillow==4.3.0
urllib3==1.22
opencv-python==3.3.0.9
IPython==5.5.0
autolab-perception==0.0.4
